Allow dynamic beatCount assignment.
Removed hardcoded beat count from multiple files to make it easier to adjust grid size. Current plan is to change grid size if the user is on mobile, but would like to make it possible for users to adjust the beatCount via a settings/config panel later on down the road.